+++ packages/codesandboxer/CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.2.0
+
+Add a new argument to `fetchFiles`, and `fetchRelativeFile` that is a config object.
+
+To the config object add `allowJSX` as a property with a boolean value.
+
+Codesandboxer can now load jsx files if you opt into it.
+
+- allow loading of JSX files
+- Add tests using fixtures in repo to test file resolution
+
 # 0.1.1
 
 Stop using * import due to struggles with transform-runtime